+++ prov/cxi/Makefile.include
@@ -26,6 +26,7 @@ _cxi_files = \
 	prov/cxi/src/cxip_rma.c \
 	prov/cxi/src/cxip_mr.c \
 	prov/cxi/src/cxip_msg.c \
+	prov/cxi/src/cxip_msg_hpc.c \
 	prov/cxi/src/cxip_atomic.c \
 	prov/cxi/src/cxip_iomm.c \
 	prov/cxi/src/cxip_faults.c \
